move rust to compositor
