+++ packages/compositor/rust/errors.rs
