+++ packages/compositor/rust/get_silent_parts.rs
