+++ packages/compositor/rust/global_printer.rs
