+++ packages/compositor/rust/max_cache_size.rs
