+++ packages/compositor/rust/opened_stream.rs
