+++ packages/docs/docs/contributing/rust.mdx
@@ -23,10 +23,9 @@ curl https://sh.rustup.rs -sSf | sh
 To build the Rust parts for your operating system, run:
 
 ```sh
+cd packages/compositor
 bun build.ts
 ```
-
-while you are in the `packages/renderer` directory.
 
 ## Building for all platforms
