--- packages/renderer/toolchains.mjs
@@ -1,6 +1,0 @@
-export const toolchains = [
-	'x86_64_gnu_toolchain',
-	'x86_64_musl_toolchain',
-	'aarch64_musl_toolchain',
-	'aarch64_gnu_toolchain',
-];
